Modify platform check for V8 build.
The platform check was resolving incorrectly on SLES 12.2. Replace it
with 'uname -m'. Note that this won't work for cross compilation but
we already use 'uname -m' elsewhere so this commit doesn't make things
any worse than they already are.

Once the version of V8 that is pulled in is updated globally this problem
will disappear.